Created ticket tests folder and moved test there.
