--- tests/manual/t-475.md
@@ -1,5 +1,0 @@
-@bender-ui: collapsed
-
-# Converting model to AttributeElement issue
-
-[ckeditor5-engine#475](https://github.com/ckeditor/ckeditor5-engine/issues/475)
