+++ tests/tickets/475/475.html
@@ -4,5 +4,5 @@
 </head>
 
 <div id="editor">
-	<p>ckeditor from http://www.cksource.com is awesome! :)</p>
+	<p>ckeditor from http://www.cksource.com is awesome!</p>
 </div>
