+++ tests/tickets/475/475.md
@@ -1,0 +1,11 @@
+@bender-ui: collapsed
+@bender-tags: ticket, 475
+
+### Auto-linking manual test [#475](https://github.com/ckeditor/ckeditor5-engine/issues/475)
+
+ - put the caret at the end of the line,
+ - type the link (http://cksource.com).
+
+Expected result: link should appear where you typed it. It should be change the `src` property as you type.
+
+Note: this is not a full feature, some other cases may not work.
